+++ ChangeLog.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 
 
 
+# Release 3.25 - 24/07/2024
+
+- FIX DA025399 : GETPOST type integer n'existe pas - *22/08/2024* - 3.25.3
+- FIX : CKeditor no check version to avoid the error message  - *20/08/2024* - 3.25.2
+- FIX : Title summary - *24/07/2024* - 3.25.1
+- FIX : Compat v20 
+  Changed Dolibarr compatibility range to 16 min - 20 max - *11/07/2024* - 3.25.0
+- Ajout du hook pdfgeneration dans la liste des hooks accepté par le module - *12/07/2024* - 3.25.0
+
 # Release 3.24 - 08/04/2024
+
+- FIX : DA024845 : Le module sous total amène des erreurs dans les sauts de page lorsque l'on arrive tout juste en bas de page. - *24/07/2024* - 3.24.5
 - FIX : Module_number missed in subTotal class. function  addSubTotalLine function  in test order_supplier - *26/06/2024* - 3.24.4  
 - FIX : display totalht ligne 'non compris' - *12/04/2024* - 3.24.3  
 - FIX : Title extrafields wasn't working - *12/04/2024* - 3.24.2  
@@ -13,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - NEW : Ajout des quantités par sous totaux sur l'interface - *27/03/2024* - 3.24.0
 
 # Release 3.23 - 18/12/2023
+
 - FIX : DA024939 - Added static method hasBreakPage to check if a line has a break page or not - *17/05/2024* - 3.23.10
 - FIX : DA024587 - Les totaux remisés sur le PDF Sponge sont erronés  - *20/03/2024* - 3.23.9
 - FIX : Suite a l'issue #379, la création de facture d'acompte avec un montant variable change les qté des lignes générées par le module donc on utilise un trigger pour remettre les bonnes qtés - *18/03/2024* - 3.23.8